errors: move QueryError to scylla crate
In addition, marked it as non_exhaustive.
