+++ scylla/src/statement/batch.rs
@@ -210,13 +210,13 @@ impl<'a: 'b, 'b> From<&'a BatchStatement>
 }
 
 pub(crate) mod batch_values {
-    use scylla_cql::errors::QueryError;
     use scylla_cql::types::serialize::batch::BatchValues;
     use scylla_cql::types::serialize::batch::BatchValuesIterator;
     use scylla_cql::types::serialize::row::RowSerializationContext;
     use scylla_cql::types::serialize::row::SerializedValues;
     use scylla_cql::types::serialize::{RowWriter, SerializationError};
 
+    use crate::errors::QueryError;
     use crate::routing::Token;
 
     use super::BatchStatement;
