+++ scylla/src/transport/connection_pool.rs
@@ -1,8 +1,8 @@
 #[cfg(feature = "cloud")]
 use crate::cloud::set_ssl_config_for_scylla_cloud_host;
 
+use crate::errors::QueryError;
 use crate::routing::{Shard, ShardCount, Sharder};
-use crate::transport::errors::QueryError;
 use crate::transport::{
     connection,
     connection::{Connection, ConnectionConfig, ErrorReceiver, VerifiedKeyspaceName},
