+++ scylla/src/transport/large_batch_statements_test.rs
@@ -1,8 +1,9 @@
 use assert_matches::assert_matches;
 
-use scylla_cql::errors::{BadQuery, QueryError};
+use scylla_cql::errors::BadQuery;
 
 use crate::batch::BatchType;
+use crate::errors::QueryError;
 use crate::query::Query;
 use crate::test_utils::setup_tracing;
 use crate::{
